requirements.txt
alembic
async-timeout
bcrypt
beautifulsoup4
blinker
bs4
certifi
charset-normalizer
click
colorama
dnspython
email-validator
Flask
Flask-Bcrypt
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
greenlet
gunicorn
html5lib
idna
importlib-metadata
importlib-resources
itsdangerous
Jinja2
Mako
MarkupSafe
packaging
python-dotenv
redis
requests
six
soupsieve
SQLAlchemy
typing_extensions
urllib3
webencodings
Werkzeug
WTForms
zipp
